The multi-wave parity port from original Xonotic - 114 commits, 928 files
(+218k/-7k). Adds gametypes, weapons, mutators, monsters/turrets/vehicles, map
objects, bot AI, warpzone portal render, HUD + client presentation, net
serialization, the machine-checkable parity registry under planning/parity, and
the load-time asset-pooling / new .mdl importer / cvar-persistence bank.

Reviewed before merge via a subsystem fan-out + verification pass; the two review
gaps (HUD/client-render deep files, and the post-bank loader/format/cvar work)
were closed and came back clean. Functional fixes landed on the branch first:
cvar-style + moveto votes, round-handler double-wiring (Invasion/KeyHunt/
Onslaught), the Buffs ammo/flight state leak, the map-vote client decode, the
dead CheckRound duplicates (FreezeTag + ClanArena), the minigame cursor latch,
the headless mipmap-load segfault guard, and the M1-M14 minor backlog.

Build clean, 2953 tests green, clean fast-forward (no conflicts).

+++ game/client/particles/ParticleRouter.cs
@@ -56,7 +56,7 @@ public sealed class ParticleRouter
     /// <paramref name="velocity"/> are Quake space; for trails <paramref name="velocity"/> is the END point.
     /// </summary>
     public bool Route(string effectName, IReadOnlyList<EffectInfoEmitter> blocks,
-        NVec3 origin, NVec3 velocity, int count, bool isTrail, Color? color)
+        NVec3 origin, NVec3 velocity, float count, bool isTrail, Color? color)
     {
         if (Faithful is null && Modern is null)
             return false;
@@ -102,7 +102,7 @@ public sealed class ParticleRouter
     }
 
     private void SpawnFaithful(string effectName, IReadOnlyList<EffectInfoEmitter> blocks,
-        EffectStyleEntry style, ModernPreset preset, NVec3 origin, NVec3 velocity, int count, bool isTrail, Color? color)
+        EffectStyleEntry style, ModernPreset preset, NVec3 origin, NVec3 velocity, float count, bool isTrail, Color? color)
     {
         // The blocks passed in are the effect's REAL effectinfo definition (Route is only called with a
         // non-empty upstream lookup) — the faithful backend always renders those verbatim, even when the
